feat: extract chat and review components
