+++ frontend/src/lib/components/ReviewForm.svelte
@@ -1,0 +1,52 @@
+<script lang="ts">
+  import { petitionData, appState, prevStep, nextStep } from '$lib/stores'
+  import { FIELD_LABELS } from '$lib/constants'
+  import { validatePetition } from '$lib/utils'
+  import { get } from 'svelte/store'
+  import type { PetitionData } from '$lib/types'
+
+  const fieldEntries: [keyof PetitionData, string][] = Object.entries(
+    FIELD_LABELS
+  ) as [keyof PetitionData, string][]
+
+  function handleReviewNext() {
+    const validation = validatePetition(get(petitionData))
+    if (validation.isValid) {
+      appState.update(s => ({ ...s, error: undefined }))
+      nextStep()
+    } else {
+      appState.update(s => ({ ...s, error: 'Please complete required fields' }))
+    }
+  }
+</script>
+
+<div class="mb-4">
+  {#each fieldEntries as [field, label]}
+    <div class="mb-2">
+      <label class="font-bold" for={`${field}-input`}>{label}</label>
+      <input
+        id={`${field}-input`}
+        class="w-full border p-2 rounded"
+        value={$petitionData[field] ?? ''}
+        on:input={(e) =>
+          petitionData.update(d => ({
+            ...d,
+            [field]: (e.target as HTMLInputElement).value
+          }))
+        }
+      />
+    </div>
+  {/each}
+</div>
+<div class="flex gap-2">
+  <button class="bg-gray-200 px-3 py-1 rounded" on:click={prevStep}>
+    Back
+  </button>
+  <button
+    class="bg-blue-500 text-white px-3 py-1 rounded"
+    on:click={handleReviewNext}
+  >
+    Generate
+  </button>
+</div>
+
